Monitor services by label instead of name
Motivation
------------
In some infrastructures it is easier and more flexible to identify a
service by labels rather than by name. This is especially true if the
name is getting a suffix via Kustomize.

Modifications
---------------
Optionally allow a list of labels used to identify the service(s) to
monitor rather than a name.

Results
-------
This change is fully backward compatible, any configuration which
supplies a name will work as-is.

If labels are supplied, more than one service may be monitored. In
this case the application is considered active if at least one service
includes an endpoint for the pod.